MicroPython: Refactor into multiple files.
* Fix poly list bug
* Add ellipse and fellipse
