+++ micropython/modules/picosystem/micropython.cmake
@@ -3,8 +3,14 @@ string(TOUPPER ${MOD_NAME} MOD_NAME_UPPER)
 add_library(usermod_${MOD_NAME} INTERFACE)
 
 target_sources(usermod_${MOD_NAME} INTERFACE
-    ${CMAKE_CURRENT_LIST_DIR}/${MOD_NAME}.c
-    ${CMAKE_CURRENT_LIST_DIR}/${MOD_NAME}.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/picosystem.c
+    ${CMAKE_CURRENT_LIST_DIR}/picosystem.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/voice.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/buffer.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/state.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/primitives.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/utility.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/hardware.cpp
     ${CMAKE_CURRENT_LIST_DIR}/../../../libraries/picosystem.cpp
     ${CMAKE_CURRENT_LIST_DIR}/../../../libraries/audio.cpp
     ${CMAKE_CURRENT_LIST_DIR}/../../../libraries/blend.cpp
@@ -35,8 +41,16 @@ set_source_files_properties(
     PROPERTIES COMPILE_FLAGS
     "-Wno-discarded-qualifiers -Wno-implicit-int"
 )
+
 set_source_files_properties(
     ${CMAKE_CURRENT_LIST_DIR}/picosystem.cpp
     PROPERTIES COMPILE_FLAGS
     "-DDYNAMIC_BUFFER=1 -DPIXEL_DOUBLE=1"
+)
+
+# TODO fix sign compare issues in codebase
+set_source_files_properties(
+    ${CMAKE_CURRENT_LIST_DIR}/../../../libraries/primitives.cpp
+    PROPERTIES COMPILE_FLAGS
+    "-Wno-error=sign-compare"
 )
